+++ tests/test_settings.py
@@ -2758,7 +2758,7 @@ def main() -> None:
     defaults = {}
     setting_ids = set()
 
-    assert data.version == "2.9.8"
+    assert data.version == "2.9.9"
 
     gradient_name = localization["mod_name"]["en"]
     assert gradient_name.startswith("{#color(174,239,105)}B")